wip(pipeline/track2): pose inputs for tcn/tcan variants.
There is something clearly wrong with the prediction outputs of the
models when provided pose inputs. If the pose data is empty, then the
model predicts with 0 issues (albeit with worse metrics than the purely
bbox ones). I suspect it's something to do with missing data in the
training set that is unexpected in the pipeline.